[IMP] mail, web, *: add support for guest users
* = crm_livechat, hr, hr_holidays, im_livechat, mail_bot, purchase, sms,
    snailmail, survey, test_discuss_full, test_mail, web_editor, website,
    website_livechat

 - Create new model `mail.guest` for guests.
 - Rewrite some RPCs to target routes rather than model methods so that
   guests are able to use them.
 - Patch JS and python models to support guests.
 - Create a stand-alone page and boot the channel in it.

task-2494829

closes odoo/odoo#75496

Related: odoo/enterprise#20417

+++ addons/mail/static/src/models/notification_group_manager/notification_group_manager.js
@@ -13,6 +13,9 @@ function factory(dependencies) {
         //----------------------------------------------------------------------
 
         computeGroups() {
+            if (!this.messaging.currentPartner) {
+                return;
+            }
             for (const group of this.groups) {
                 group.delete();
             }
